Se modificó main
Se modificó main
Se creó home_page dentro de src
