+++ demotic/lib/src/pages/home_page.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class HomePage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(  
+      
+      appBar: AppBar(
+        title: Text('Home'),
+        backgroundColor: Colors.teal,
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Text('Home', style: TextStyle( fontSize: 30.0, fontWeight: FontWeight.bold),)
+          ],
+        ),
+      ),
+    );
+  }
+}
